+++ account-kit/react/src/components/auth/card/error/connection-error.tsx
@@ -7,7 +7,8 @@ import { EOAWallets, type ConnectionErrorProps } from "./types.js";
 import { WalletIcon } from "./icons/wallet-icon.js";
 import { OAuthConnectionFailed } from "../../../../icons/oauth.js";
 import { capitalize } from "../../../../utils.js";
-import { useSigner } from "../../../../hooks/useSigner.js";
+import { disconnect } from "@account-kit/core";
+import { useAlchemyAccountContext } from "../../../../context.js";
 
 export const walletTypeConfig = [
   { name: "Coinbase Wallet", key: EOAWallets.COINBASE_WALLET },
@@ -23,21 +24,12 @@ export const ConnectionError = ({
   handleUseAnotherMethod,
   customErrorMessage,
 }: ConnectionErrorProps) => {
-  const signer = useSigner();
+  const { config } = useAlchemyAccountContext();
 
   useEffect(() => {
     // Terminate any inflight authentication on Error...
-    if (
-      connectionType === "wallet" &&
-      EOAConnector === EOAWallets.WALLET_CONNECT
-    ) {
-      clearWalletConnectStore();
-    }
-
-    if (signer) {
-      signer.disconnect();
-    }
-  }, [signer, connectionType, EOAConnector]);
+    disconnect(config);
+  }, [config]);
 
   const getHeadingText = useMemo(() => {
     if (customErrorMessage) {
@@ -112,24 +104,4 @@ export const ConnectionError = ({
       </Button>
     </div>
   );
-};
-
-const clearWalletConnectStore = () => {
-  // Open Wallet Connect Indexed DB
-  const dbOpenRequest = indexedDB.open("WALLET_CONNECT_V2_INDEXED_DB");
-
-  dbOpenRequest.onsuccess = () => {
-    const db = dbOpenRequest.result;
-
-    const txn = db.transaction(["keyvaluestorage"], "readwrite");
-
-    const store = txn.objectStore("keyvaluestorage");
-
-    // Clear Store
-    store.clear();
-  };
-
-  dbOpenRequest.onerror = () => {
-    console.error("Error opening Wallet Connect DB. Cannot clear store.");
-  };
 };